Replace Web stop/start with stopPIA/startPIA rather than psadmin.
psadmin -w commands will not run from a non-interactive shell. This
breaks Ansible, Puppet and systemd using pscontrol for web servers.
I tried workaround including reasonalbe and unreasonable. Oracle does
not consider this to be a bug according to what I found on support, but
there is reference to it being changed in 8.56 possibly... just upgrade!